requirements.txt
selenium
webdriver-manager
opencv-python
pytube
Pillow
PyYAML